requirements.txt
numpy
PyOpenGL
glfw
PyOpenGL-accelerate